cambios xd

+++ Controllers/ReservationController.cs
@@ -37,6 +37,27 @@ namespace ApiHoteleria.Controllers
         }
 
 
+        private string getPersonId(string token)
+        {
+
+            var handler = new JwtSecurityTokenHandler();
+            var jsonToken = handler.ReadToken(token);
+            var tokenS = jsonToken as JwtSecurityToken;
+            var claims = tokenS.Claims.Select(claim => (claim.Type, claim.Value)).ToList();
+            string userId = "";
+            for (int i = 0; i < claims.Count; i++)
+            {
+                System.Diagnostics.Debug.WriteLine("EL TYPE DEL TOKEN ES " + claims[i].Type);
+                if (claims[i].Type == "jti")
+                {
+                    userId = claims[i].Value;
+                }
+            }
+            System.Diagnostics.Debug.WriteLine("EL PERSON DEL TOKEN ES " + userId);
+            return userId;
+        }
+
+
         [Authorize]
         [HttpGet]
         [Route("getReservationsDetails")]
@@ -124,19 +145,32 @@ namespace ApiHoteleria.Controllers
                     return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
                 }
 
-                var existingRoom = connection.Query<Reservation>("SELECT r.Status FROM reservation r" +
-                    "INNER JOIN room rm ON rm.Hotel_ID = r.Hotel_ID" +
-                    "WHERE r.Status = 'Pendiente' AND r.Client_ID = @clientId AND rm.Room_ID = @roomId",new { clientId, roomId = hotel.Room_ID}).ToList();
+                string personId = getPersonId(authorization.ToString().Replace("Bearer ", ""));
 
-                if(existingRoom.Count > 0)
+                if (personId == null)
+                {
+                    statusCode = (int)HttpStatusCode.Forbidden;
+                    message = "Invalid token";
+                    return StatusCode((int)HttpStatusCode.Forbidden, new { statusCode, message });
+                }
+
+
+                var existingRoom = connection.Query<Reservation>("SELECT r.Status FROM reservation r " +
+                    "INNER JOIN room rm ON rm.Hotel_ID = r.Hotel_ID " +
+                    "WHERE r.Status = 'Pendiente' AND r.Client_ID = @clientId AND rm.Room_ID = @roomId",new { clientId=personId, roomId = hotel.Room_ID}).ToList();
+
+                if(existingRoom.Count > 0&& existingRoom[0].Status != "Pendiente")
                 {
                     message = "Room already reserved";
                     statusCode = (int)HttpStatusCode.Forbidden;
                     return StatusCode((int)HttpStatusCode.Forbidden, new { statusCode, message });
                 }
 
+
+
+
                 var userReservation = connection.Query<Reservation>("SELECT * FROM reservation WHERE Client_ID = @client_id AND Status = 'Pendiente'", 
-                    new { client_id = clientId }).ToList();
+                    new { client_id = personId }).ToList();
 
                 DateTime fechaUno = Convert.ToDateTime(hotel.Check_In_Date).Date;
                 DateTime fechaDos = Convert.ToDateTime(hotel.Check_Out_Date).Date;
@@ -154,7 +188,7 @@ namespace ApiHoteleria.Controllers
 
                     connection.Execute("INSERT INTO reservation (Hotel_ID, Client_ID, Total_Cost, Status) " +
                         "VALUES (@hotel_id, @client_id, @total_cost, 'Pendiente')",
-                    new { hotel.Hotel_ID, client_id = clientId, total_cost = totalCost });
+                    new { hotel.Hotel_ID, client_id = personId, total_cost = totalCost });
                     int userId = connection.QuerySingle<int>("SELECT Reservation_ID FROM reservation order by Reservation_ID DESC LIMIT 1");
 
                     reservationId = userId;
@@ -173,7 +207,7 @@ namespace ApiHoteleria.Controllers
                     "@check_in_date, @check_out_date, @room_id)",
                                        new { reservation_id = reservationId, hotel.Check_In_Date, hotel.Check_Out_Date, hotel.Room_ID });
 
-                response = Ok(new { statusCode, message });
+                response = Ok(new { statusCode, message, reservationId});
 
                 return response;
             }
@@ -230,8 +264,17 @@ namespace ApiHoteleria.Controllers
                     return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
                 }
 
+                string personId = getPersonId(authorization.ToString().Replace("Bearer ", ""));
+
+                if (personId == null)
+                {
+                    statusCode = (int)HttpStatusCode.Forbidden;
+                    message = "Invalid token";
+                    return StatusCode((int)HttpStatusCode.Forbidden, new { statusCode, message });
+                }
+
                 var userReservation = connection.Query<Reservation>("SELECT * FROM reservation WHERE Client_ID = @client_id AND Status = 'Pendiente'",
-                    new { client_id = clientId }).ToList();
+                    new { client_id = personId }).ToList();
 
                 if (userReservation == null || userReservation.Count == 0)
                 {
@@ -240,12 +283,11 @@ namespace ApiHoteleria.Controllers
                     return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
                 }
 
-
-                var availableRooms = connection.Query<Rooms>("SELECT rd.Room_ID, r.Room_Number, r.Status  FROM reservation_detail rd " +
+                var availableRooms = connection.Query<Rooms>("SELECT rd.Check_In_Date, rd.Check_Out_Date, rd.Room_ID, r.Room_Number, r.Status  FROM reservation_detail rd " +
                     "INNER JOIN room r ON r.Room_ID = rd.Room_ID " +
                     "INNER JOIN reservation rv ON rv.Reservation_ID = rd.Reservation_ID " +
                     "INNER JOIN room_type rt ON rt.Type_ID = r.Type_ID  WHERE rv.Client_ID " +
-                    "= @user_id AND rv.Reservation_ID = @reservation_id", new { user_id = clientId, reservation_id = userReservation[0].Reservation_ID }).ToList();
+                    "= @user_id AND rv.Reservation_ID = @reservation_id", new { user_id = personId, reservation_id = userReservation[0].Reservation_ID }).ToList();
 
 
                 if (availableRooms == null || availableRooms.Count == 0)
@@ -255,9 +297,31 @@ namespace ApiHoteleria.Controllers
                     return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
                 }
 
+                var availableRoomDetails = connection.Query<Rooms>("select r.Room_Number, rd.Check_In_Date, rd.Check_Out_Date FROM room r " +
+                    "INNER JOIN reservation_detail rd ON r.Room_ID = r.Room_ID and r.Status = 'Confirmada'").ToList();
+                if (availableRoomDetails.Count > 0)
+                {
+                    for (int i = 0; i < availableRoomDetails.Count; i++)
+                    {
+                        DateTime checkInDate = availableRoomDetails[i].Check_In_Date;
+                        DateTime checkOutDate = availableRoomDetails[i].Check_Out_Date;
+
+                        foreach (var room in availableRooms)
+                        {
+                            if ((checkInDate < room.Check_Out_Date && checkOutDate > room.Check_In_Date))
+                            {
+                                statusCode = (int)HttpStatusCode.NotFound;
+                                message = "Room " + availableRoomDetails[i].Room_Number + " not available";
+                                return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
+                            }
+                        }
+                    }
+                }
+
+
                 for (int i = 0; i < availableRooms.Count; i++)
                 {
-                    if (availableRooms[i].Status != "Disponible")
+                    if (availableRooms[i].Status != "Disponible"  )
                     {
                         statusCode = (int)HttpStatusCode.NotFound;
                         message = "Room " + availableRooms[i].Room_Number + " not available";
@@ -388,19 +452,50 @@ namespace ApiHoteleria.Controllers
                     return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
                 }
 
-                IActionResult response = Unauthorized();
 
-                if (String.IsNullOrEmpty(room_id.ToString()))
+                string personId = getPersonId(authorization.ToString().Replace("Bearer ", ""));
+
+                if (personId == null)
                 {
                     statusCode = (int)HttpStatusCode.Forbidden;
-                    message = "Incomplete request";
+                    message = "Invalid token";
                     return StatusCode((int)HttpStatusCode.Forbidden, new { statusCode, message });
                 }
 
-                connection.Execute("DELETE FROM reservation_detail WHERE Reservation_Detail_ID = @room_id", new { room_id });
+  
+                var currentReservation = connection.Query<Reservation>("SELECT * FROM reservation WHERE Client_ID = @personId " +
+                    "AND Status = 'Pendiente'", new { personId}).ToList();
 
-                response = Ok(new { statusCode, message });
+                if (currentReservation == null || currentReservation.Count == 0)
+                {
+                    statusCode = (int)HttpStatusCode.NotFound;
+                    message = "Person not found";
+                    return StatusCode((int)HttpStatusCode.NotFound, new { statusCode, message });
+                }
 
+                var room = connection.Query<Rooms>("SELECT r.*, rt.Capacity, rt.Price_Per_Night, rt.Description as type FROM room r INNER JOIN room_type rt " +
+                    "ON rt.Type_ID = r.Type_ID AND Room_ID=@room_id", new { room_id }).ToList();
+
+                var detail = connection.Query<Details>("SELECT * FROM reservation_detail WHERE Reservation_ID = @currentReservation", 
+                    new { currentReservation = currentReservation[0].Reservation_ID }).ToList();
+
+                DateTime fechaUno = Convert.ToDateTime(detail[0].Check_In_Date).Date;
+                DateTime fechaDos = Convert.ToDateTime(detail[0].Check_Out_Date).Date;
+                TimeSpan difFechas = fechaDos - fechaUno;
+                int days = (int)difFechas.TotalDays;
+                string dias = Convert.ToString(days);
+                double totalCost = room[0].Price_Per_Night * int.Parse(dias);
+                double totalReservation = currentReservation[0].Total_Cost - totalCost;
+
+                System.Diagnostics.Debug.WriteLine(totalReservation);
+                System.Diagnostics.Debug.WriteLine(currentReservation[0].Reservation_ID);
+
+                connection.Execute("DELETE FROM reservation_detail WHERE Reservation_ID = @currentReservation AND Room_ID=@room_id",
+                    new { currentReservation = currentReservation[0].Reservation_ID, room_id });
+                connection.Execute("UPDATE reservation rv SET Total_Cost = @totalReservation WHERE rv.Reservation_ID = @currentReservation",
+                    new { totalReservation, currentReservation = currentReservation[0].Reservation_ID });
+
+                IActionResult response = Ok(new { statusCode, message });
                 return response;
             }
             catch (Exception e)
@@ -449,6 +544,16 @@ namespace ApiHoteleria.Controllers
 
                 var reservation_detail = new List<Reservation>();
                 System.Diagnostics.Debug.WriteLine("ID ES " + client_id);
+
+                string personId = getPersonId(authorization.ToString().Replace("Bearer ", ""));
+
+                if (personId == null)
+                {
+                    statusCode = (int)HttpStatusCode.Forbidden;
+                    message = "Invalid token";
+                    return StatusCode((int)HttpStatusCode.Forbidden, new { statusCode, message });
+                }
+
 
                 if (client_id != 0)
                 {
@@ -549,6 +654,54 @@ namespace ApiHoteleria.Controllers
                 return StatusCode((int)HttpStatusCode.InternalServerError, new { statusCode, message });
             }
 
+        }
+
+
+        [Authorize]
+        [HttpGet]
+        [Route("getExistingReservation")]
+        public IActionResult getExistingReservation([FromQuery] int user_id,[FromServices] MySqlConnection connection)
+        {
+            try
+            {
+                IActionResult response = Unauthorized();
+                var reservatiosnResponse = connection.Query<Reservation>("SELECT * FROM reservation WHERE Client_ID = @clientId " +
+                    "AND Status = 'Pendiente'", new {clientId = user_id}).ToList();
+
+                if(reservatiosnResponse.Count > 0)
+                {
+                    var rooms = connection.Query<Rooms>("SELECT * FROM room r INNER JOIN reservation_detail rd " +
+                        "ON rd.Room_ID = r.Room_ID INNER JOIN reservation rs " +
+                        "ON rs.Reservation_ID = rd.Reservation_ID WHERE rs.Client_ID = @clientId " +
+                        "AND rs.Reservation_ID = @reservationId", new {clientId = user_id, reservationId = reservatiosnResponse[0].Reservation_ID }).ToList();
+
+
+                    return Ok(new
+                    {
+                        statusCode = (int)HttpStatusCode.OK,
+                        data = reservatiosnResponse,
+                        rooms
+                    });
+
+                }
+
+
+                return Ok(new
+                {
+                    statusCode = (int)HttpStatusCode.OK,
+                    data = reservatiosnResponse
+                });
+            }
+            catch (Exception e)
+            {
+                System.Diagnostics.Debug.WriteLine(e);
+                return BadRequest(
+                    new
+                    {
+                        statusCode = (int)HttpStatusCode.BadRequest,
+                        message = "An error has ocurred: " + e.Message
+                    });
+            }
         }
